Update TableExporter to prepare values so individual tasks don't have to
